Move check wrapper to a better location
Since it is used internally by the collector, make it an internal
subpackage of the collector.
