+++ pkg/collector/collector.go
@@ -13,6 +13,7 @@ import (
 	"go.uber.org/atomic"
 
 	"github.com/DataDog/datadog-agent/pkg/collector/check"
+	"github.com/DataDog/datadog-agent/pkg/collector/internal/middleware"
 	"github.com/DataDog/datadog-agent/pkg/collector/runner"
 	"github.com/DataDog/datadog-agent/pkg/collector/runner/expvars"
 	"github.com/DataDog/datadog-agent/pkg/collector/scheduler"
@@ -36,7 +37,7 @@ type Collector struct {
 
 	scheduler *scheduler.Scheduler
 	runner    *runner.Runner
-	checks    map[check.ID]*scheduler.CheckWrapper
+	checks    map[check.ID]*middleware.CheckWrapper
 
 	m sync.RWMutex
 }
@@ -44,7 +45,7 @@ type Collector struct {
 // NewCollector create a Collector instance and sets up the Python Environment
 func NewCollector(paths ...string) *Collector {
 	c := &Collector{
-		checks:         make(map[check.ID]*scheduler.CheckWrapper),
+		checks:         make(map[check.ID]*middleware.CheckWrapper),
 		state:          atomic.NewUint32(stopped),
 		checkInstances: int64(0),
 	}
@@ -113,7 +114,7 @@ func (c *Collector) RunCheck(inner check.Check) (check.ID, error) {
 	c.m.Lock()
 	defer c.m.Unlock()
 
-	ch := scheduler.NewCheckWrapper(inner)
+	ch := middleware.NewCheckWrapper(inner)
 
 	var emptyID check.ID
